file 文件库相关api

+++ applications/file.py
@@ -13,7 +13,8 @@ from model.user import User
 from common.jwt import get_current_user_authorizer
 from common.mongodb import AsyncIOMotorClient, get_database
 
-from crud.file import create_file, get_file_list, count_file_by_query, delete_file
+from crud.file import create_file, get_file_list, count_file_by_query, delete_file, update_file, get_file
+from crud.work_history import count_work_history_by_query
 from model.file import FileCreateModel
 from common.upload import MinioUploadPrivate
 from common.common import contenttomd5, tokenize_words
@@ -22,7 +23,7 @@ router = APIRouter()
 _platform = platform.system().lower()
 
 
-@router.post('/upload/file', tags=['upload'], name='文档上传')
+@router.post('/file', tags=['file'], name='文档上传')
 async def upload_file(file: UploadFile = File(...), user: User = Depends(get_current_user_authorizer()),
                       db: AsyncIOMotorClient = Depends(get_database)):
     attr = file.filename.rsplit('.')[-1]
@@ -82,6 +83,7 @@ async def upload_file(file: UploadFile = File(...), user: User = Depends(get_cur
         logger.info('未分词文档，自动分词中')
         parsed_content = tokenize_words(origin_content)
         logger.info(parsed_content)
+        data.is_check = True
     else:
         parsed_content = origin_content
     # 提交分词结果
@@ -93,10 +95,10 @@ async def upload_file(file: UploadFile = File(...), user: User = Depends(get_cur
     return {'id': data.id}
 
 
-@router.get('/my/file', tags=['user'], name='我的文件')
-async def get_file(user_id: str = None, search: str = None, page: int = 1, limit: int = 20,
-                   user: User = Depends(get_current_user_authorizer()),
-                   db: AsyncIOMotorClient = Depends(get_database)):
+@router.get('/my/file', tags=['file'], name='我的文件')
+async def get_my_file(user_id: str = None, search: str = None, is_check: bool = None, page: int = 1, limit: int = 20,
+                      user: User = Depends(get_current_user_authorizer()),
+                      db: AsyncIOMotorClient = Depends(get_database)):
     if 0 in user.role:
         u_id = user_id
     else:
@@ -104,6 +106,8 @@ async def get_file(user_id: str = None, search: str = None, page: int = 1, limit
     query_obj = {'user_id': u_id}
     if search is not None:
         query_obj['file_name'] = {'$regex': search}
+    if is_check is not None:
+        query_obj['is_check'] = is_check
     data = await get_file_list(db, query_obj, page=page, limit=limit)
     count = await count_file_by_query(db, query_obj)
     return {
@@ -112,23 +116,59 @@ async def get_file(user_id: str = None, search: str = None, page: int = 1, limit
     }
 
 
-@router.patch('/file', tags=['user'], name='修改文件（parsed）')
-async def patch_file(file_id: str,
+@router.get('/file/content', tags=['file'], name='某一文件内容')
+async def get_file_content(file_id: str,
+                           user: User = Depends(get_current_user_authorizer()),
+                           db: AsyncIOMotorClient = Depends(get_database)):
+    db_file = await get_file(db, {'id': file_id})
+    if not db_file:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '文件不存在')
+    if db_file.user_id != user.id and 0 not in user.role:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '权限不足')
+    m = MinioUploadPrivate()
+    content = m.get_object(db_file.parsed)
+    return {'data': content}
+
+
+@router.patch('/file', tags=['file'], name='修改文件（parsed,is_check）')
+async def patch_file(file_id: str = Body(...), content: str = Body(None), is_check: bool = Body(None),
                      user: User = Depends(get_current_user_authorizer()),
                      db: AsyncIOMotorClient = Depends(get_database)):
-    pass
+    # 1.内容更新到minio，2.更新file的p_hash
+    db_file = await get_file(db, {'id': file_id})
+    if not db_file:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '文件不存在')
+    if db_file.user_id != user.id and 0 not in user.role:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '权限不足')
+    update_obj = {}
+    if content is not None:
+        m = MinioUploadPrivate()
+        # 上传文件
+        m.commit(content.encode('utf-8'), db_file.parsed)
+        new_hash = contenttomd5(content.encode('utf-8'))
+        update_obj['p_hash'] = new_hash
+    if is_check is not None:
+        update_obj['is_check'] = is_check
+    await update_file(db, {'id': file_id}, {'$set': update_obj})
+    return {'msg': '2002'}
 
 
-@router.delete('/file', tags=['user'], name='删除文件')
+@router.delete('/file', tags=['file'], name='删除文件')
 async def del_file(file_id: str,
                    user: User = Depends(get_current_user_authorizer()),
                    db: AsyncIOMotorClient = Depends(get_database)):
-    # todo 文件不能在work_history中出现
+    db_file = await get_file(db, {'id': file_id})
+    if not db_file:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '文件不存在')
+    if db_file.user_id != user.id and 0 not in user.role:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '权限不足')
+    # 文件不能在work_history中出现
+    use_count = await count_work_history_by_query(db, {'file_id': file_id})
+    if use_count > 0:
+        raise HTTPException(HTTP_400_BAD_REQUEST, '此文件不可删除')
+    m = MinioUploadPrivate()
+    m.remove(db_file.parsed)
+    m.remove(db_file.origin)
     await delete_file(db, {'id': id, 'user_id': user.id})
+    return {'msg': '2002'}
 
-
-# TODO
-'''
-1.词频统计词库导入（管理员）
-2.已有词典导入（管理员）
-'''
